add chapter server functions to actions file
